two configs
