+++ assets/config.js
@@ -1,17 +1,31 @@
-module.exports = {
-    assetHost: 'https://jason.now.sh',
+const config = {
     flow: 'a4f64f1b-85bc-477e-b706-de313a022979',
     languages: {
         eng: 'English',
         spa: 'Spanish',
         fre: 'French'
-    },
-    endpoints: {
-        flows: '/assets/flows.json',
-        groups: '/assets/groups.json',
-        contacts: '/assets/contacts.json',
-        fields: '/assets/fields.json',
-        activity: '',
-        engine: ''
     }
 };
+
+module.exports =
+    process.env.NODE_ENV === 'production'
+        ? Object.assign({}, config, {
+              endpoints: {
+                  flows: 'flows',
+                  groups: 'groups',
+                  contacts: 'contacts',
+                  fields: 'fields',
+                  activity: '',
+                  engine: ''
+              }
+          })
+        : Object.assign({}, config, {
+              endpoints: {
+                  flows: '/assets/flows.json',
+                  groups: '/assets/groups.json',
+                  contacts: '/assets/contacts.json',
+                  fields: '/assets/fields.json',
+                  activity: '',
+                  engine: ''
+              }
+          });
