+++ preview/src/app.tsx
@@ -5,27 +5,11 @@ import 'raf/polyfill';
 import * as React from 'react';
 import { render } from 'react-dom';
 import { AppContainer as HotContainer } from 'react-hot-loader';
+import * as config from '../../assets/config';
 import FlowEditor from '../../src/';
 import { FlowEditorConfig } from '../../src/flowTypes';
 
 const root = document.getElementById('flow-editor');
-
-const config = {
-    flow: 'a4f64f1b-85bc-477e-b706-de313a022979',
-    languages: {
-        eng: 'English',
-        spa: 'Spanish',
-        fre: 'French'
-    },
-    endpoints: {
-        flows: 'flows',
-        groups: 'groups',
-        contacts: 'contacts',
-        fields: 'fields',
-        activity: '',
-        engine: ''
-    }
-};
 
 const renderHot = (App: React.SFC<{ config: FlowEditorConfig }>) =>
     render(
